pyproject.toml
[project]
name = "chat-bot-stream"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
